aggregatorService/requirements.txt
datetime
psycopg2-binary
boto3
pandas
mysql-connector-python